refactoring: move allocations from mixins to store getters
