+++ src/store/addon.js
@@ -1,4 +1,5 @@
 import { Addon } from "@wealthica/wealthica.js";
+import { LABELS } from "@/constants";
 
 export default {
   state: {
@@ -18,7 +19,43 @@ export default {
     currency: (state, getters) => getters.addonOptions.currency,
     currencies: state => state.currencies,
     preferredCurrencyId: state => state.preferredCurrencyId,
-    preferredCurrency: state => state.currencies[state.preferredCurrencyId]
+    preferredCurrency: state => state.currencies[state.preferredCurrencyId],
+    allocations: (state, getters) => {
+      let netAmount = 0;
+      const allocations = [0, 1].map(index => ({
+        label: LABELS[index],
+        amount: 0,
+        cntHoldings: 0,
+        percent: 0,
+        symbols: []
+      }));
+
+      getters.positions.forEach(position => {
+        const index = position.class === "equity" ? 1 : 0;
+
+        allocations[index].amount += position.value;
+        allocations[index].cntHoldings += 1;
+
+        if (position.security) {
+          allocations[index].symbols.push(
+            position.security.symbol || position.security.name
+          );
+        }
+
+        netAmount += position.value;
+      });
+
+      if (netAmount > 0) {
+        allocations.forEach((allocation, index) => {
+          allocations[index].percent = parseFloat(
+            ((100 * allocation.amount) / netAmount).toFixed(2)
+          );
+          allocations[index].symbolsList = allocation.symbols.join(", ");
+        });
+      }
+
+      return allocations;
+    }
   },
   mutations: {
     SET_ADDON(state, data) {
